requirements.txt
wheel
dj-database-url==0.4.2
Django==1.10.5
gunicorn==19.6.0
psycopg2==2.6.2
python-decouple==3.0
django-extensions==1.7.5
djangorestframework==3.5.3
django-cors-headers==2.0.0
django-filter==1.0.1
django-crispy-forms==1.6.1
Pillow==4.0.0
django-import-export==0.5.1
pytz
